+++ src/module.test.ts
@@ -1,18 +1,9 @@
-// import { standardEditorsRegistry } from '@grafana/data';
-// import { plugin } from './module';
+import { PanelPlugin } from '@grafana/data';
+import { plugin } from './module';
 
-// describe('button plugin', () => {
-//   test('editor options', () => {
-//     standardEditorsRegistry.setInit(() => {
-//       return [{ id: 'radio' }] as any;
-//     });
-//     expect(plugin.fieldConfigRegistry).toBeDefined();
-//     expect(plugin.fieldConfigRegistry.list()).toHaveLength(2);
-//   });
-// });
-
-test('fixme', () => {
-  // FIXME: there seems to be a problem with the package @grafana/data:
-  // as soon as you try to import anything from the package, the following error is thrown:
-  // LanguageProvider cannot reassign to a class (same with DataSourceApi)
+describe('button plugin', () => {
+  test('is a panel plugin with options', () => {
+    expect(plugin).toBeInstanceOf(PanelPlugin);
+    expect(plugin.panel).toBeDefined();
+  });
 });